+++ MagickWand/magick-cli.c
@@ -611,7 +611,7 @@ WandExport int ProcessCommandOptions(MagickCLI *cli_wand,int argc,char **argv,
      order of location format escapes: filename, line, column */
   cli_wand->location="at %s arg %u";
   cli_wand->filename="CLI";
-  cli_wand->line=index;  /* note first argument we will process */
+  cli_wand->line=(size_t) index;  /* note first argument we will process */
 
   if (cli_wand->wand.debug != MagickFalse)
     (void) CLILogEvent(cli_wand,CommandEvent,GetMagickModule(),
@@ -629,7 +629,7 @@ WandExport int ProcessCommandOptions(MagickCLI *cli_wand,int argc,char **argv,
     do { /* use break to loop to exception handler and loop */
 
       option=argv[i];
-      cli_wand->line=i;  /* note the argument for this option */
+      cli_wand->line=(size_t) i;  /* note the argument for this option */
 
       /* get option, its argument count, and option type */
       cli_wand->command = GetCommandOptionInfo(argv[i]);
@@ -727,7 +727,7 @@ RestoreMSCWarning
      Implicit Write of images to final CLI argument
   */
   option=argv[i];
-  cli_wand->line=i;
+  cli_wand->line=(size_t) i;
 
   /* check that stacks are empty - or cause exception */
   if (cli_wand->image_list_stack != (CLIStack *) NULL)
